Rename validUserProfile to userProfileValidators and relocate for better structure
